Add asl3-update-astdb
To help support web applications that are dependent on the "astdb.txt" file we
copy a modified version of the ASL3 "astdb.php" command, a small bit of polish,
and add a systemd .service/.timer to keep the file updated.

To minimize the impact on the sytem and network when these applications are NOT
installed we start the .service/.timer off as disabled.  When "enabled", the
system will download a copy of "astdb.txt" to the "/var/lib/asterisk" directory
and add symlinks for the web applications.